requirements.txt
pycryptodome==3.22.0
aiohttp==3.11.15
aiopg==1.4.0
python-dotenv==1.1.1